Rename RPC route files to match endpoints (#3837)
* Rename RPC route files to match endpoints

* Try fix fixtures

* Rename another fixture
