+++ ironfish/src/rpc/routes/chain/index.ts
@@ -5,8 +5,8 @@
 export * from './broadcastTransaction'
 export * from './estimateFeeRate'
 export * from './estimateFeeRates'
-export * from './exportChain'
-export * from './followChain'
+export * from './exportChainStream'
+export * from './followChainStream'
 export * from './getBlock'
 export * from './getChainInfo'
 export * from './getDifficulty'
