+++ ironfish/src/rpc/routes/index.ts
@@ -4,12 +4,12 @@
 export * from './wallet'
 export * from './config'
 export * from './chain'
-export * from './events'
+export * from './event'
 export * from './node'
-export * from './peers'
+export * from './peer'
 export * from './router'
 export * from './rpc'
-export * from './mining'
+export * from './miner'
 export * from './faucet'
-export * from './workers'
+export * from './worker'
 export * from './mempool'
