+++ ironfish/src/rpc/routes/miner/index.ts
