+++ ironfish/src/rpc/routes/peer/getBannedPeers.ts
