+++ ironfish/src/rpc/routes/wallet/index.ts
@@ -11,17 +11,17 @@ export * from './getAssets'
 export * from './getBalance'
 export * from './getBalances'
 export * from './getDefaultAccount'
-export * from './getNotes'
+export * from './getAccountNotesStream'
 export * from './getPublicKey'
-export * from './getStatus'
-export * from './getTransaction'
-export * from './getTransactions'
+export * from './getAccountsStatus'
+export * from './getAccountTransaction'
+export * from './getAccountTransactions'
 export * from './importAccount'
 export * from './mintAsset'
 export * from './postTransaction'
-export * from './removeAccount'
+export * from './remove'
 export * from './rescanAccount'
 export * from './sendTransaction'
 export * from './sendTransaction'
-export * from './useAccount'
+export * from './use'
 export * from './createTransaction'
