+++ ironfish/src/rpc/routes/wallet/remove.ts
