+++ ironfish/src/rpc/routes/worker/getStatus.ts
